Fix issue where the table view pan gesture was being disabled along with all other table view gestures when a cell was swiped. Close #21.
This wasn't an issue if the follow-up tap occurred on the non-swiped cell because the hitTest/pointInside would trigger the hide and reenable the table view gestures.  This fix always allows the table view pan gesture to continue working while still disabling the remaining gestures, mainly to disable 3D touch.