Improve Dark Mode support on iOS 13
When running on iOS 13 in Dark Mode, the background color of the SwipeActionsView was using light gray. Instead the background color should use the new .systemGroupedBackground color.

The change consists of using the new .systemGroupedBackground color for the background color on iOS 13, with a fallback to the old hardcoded color on older iOS versions.