Add shouldDelete*At methods to swipe cell delegates
Currently both `SwipeCollectionViewCell` and `SwipeTableViewCell`
automatically delete the cells on
`swipeController(:didDeleteSwipeableAt:)` and don't allow any
customization, like what's allowed in other similar methods.

While on most scenarios the default behavior is desired, there are some
scenarios where users might want to control how the deletion is
effectively made on the container (CollectionView or TableView). Some
examples are:
- Data-driven setups that use diffing
- Customize the deletion animation (table view)

To achieve that, a new delegate method was added to both the
`SwipeCollectionViewCellDelegate` and `SwipeTableViewCellDelegate` to
allow customizing the deletion on the container, having the current
implementations as the default in protocol extensions.

## Changes

- Add new `collectionView(:shouldDeleteItemAt:)` delegate method to
`SwipeCollectionViewCellDelegate`, and invoke it on
`SwipeCollectionViewCell.swipeController(:didDeleteSwipeableAt:)`.

- Add new `tableView(:shouldDeleteRowAt:)` delegate method to
`SwipeTableViewCellDelegate` and invoke it on
`SwipeTableViewCell.swipeController(:didDeleteSwipeableAt:)`.