Move `didDeleteSwipeableAt` delegate call to swipe animation completion
Following the previous changes to the cell deletion flow on swipe
completion, the actual cell deletion trigger was moved to the animation
completion so that we wait for the swipe animation to complete before
actually triggering the cell removal.

To prevent visual artifacts, the reset calls in the animation
completion were moved to the `prepareForReuse` of each cell, achieving
the same result.

## Changes

- Invoke `swipeController(:didDeleteSwipeableAt:)` delegate call on
delete animation completion

- Remove reset logic from `SwipeController`'s delete animation
completion closure, moving it to cell's `prepareForReuse`.